feat: 실제 API 연동과 화면 검증 보완
탐색·추천·영상 일정 API 계약과 비용 타입을 반영하고 알림 안내, 현재 위치, 작은 화면 스크롤, 제출 중 입력 경합을 수정한다.

236개 테스트와 Android release 및 iOS Simulator 컴파일을 통과했다. 실제 Android E2E 결과와 서버 분석 FAILED·FCM 외부 설정 등 남은 제한을 문서화한다.

SCC 기존 기능 UID를 보존해 명세 동기화를 복구하고 검증한 시트 스크롤 상태를 갱신한다.

+++ domain/src/commonMain/kotlin/com/linkit/company/domain/model/map/TripPlanMapData.kt
@@ -12,8 +12,8 @@ data class TripPlanMapData(
     /** 지도에 표시할 수 있는 장소가 없으면 null이다. */
     val center: GeoCoordinate?,
     val analysisSummary: String? = null,
-    val estimatedMinCost: Int? = null,
-    val estimatedMaxCost: Int? = null,
+    val estimatedMinCost: Long? = null,
+    val estimatedMaxCost: Long? = null,
     val costBasis: CostBasis? = null,
     val thumbnailUrl: String? = null,
 )
